pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "medical-rag"
dynamic = []
version = "0.1"
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "U.N. Owen", email = "void@some.where" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
  "faiss-cpu",
  "numpy",
  "pandas",
  "sentence_transformers",
  "torch",
  "transformers"
]

[project.urls]
Documentation = "https://github.com/U.N. Owen/medical-rag#readme"
Issues = "https://github.com/U.N. Owen/medical-rag/issues"
Source = "https://github.com/U.N. Owen/medical-rag"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/medrag tests}"

[tool.coverage.run]
source_pkgs = ["medrag", "tests"]
branch = true
parallel = true
omit = []

[tool.hatch.build.targets.wheel]
packages = ["src/medrag"]

[tool.coverage.paths]
medrag = ["src"]
tests = ["tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
